Removing async_trait and refactoring
