+++ .travis.yml
@@ -12,10 +12,8 @@ script:
   - cargo test --tests --examples
   - cargo test --doc
   - cargo test --all-features
-  # Futures 0.3 with reqwest 0.10
-  - cargo test --tests --examples --features futures-03,reqwest-010 --no-default-features
-  # Futures 0.3 without reqwest (examples will not build)
-  - cargo test --tests --features futures-03 --no-default-features
+  # Curl without reqwest (examples will not build)
+  - cargo test --tests --features curl --no-default-features
   - cargo audit
 notifications:
   email:
