+++ src/reqwest.rs
@@ -22,34 +22,25 @@ where
     Other(String),
 }
 
-#[cfg(feature = "reqwest-010")]
 pub use blocking::http_client;
 ///
 /// Error type returned by failed reqwest blocking HTTP requests.
-/// Requires "reqwest-010" feature.
 ///
-#[cfg(feature = "reqwest-010")]
 pub type HttpClientError = Error<blocking::reqwest::Error>;
 
-#[cfg(all(feature = "futures-03", feature = "reqwest-010"))]
 pub use async_client::async_http_client;
+
 ///
 /// Error type returned by failed reqwest async HTTP requests.
-/// Requires "futures-03" and "reqwest-010" feature.
 ///
-#[cfg(all(feature = "futures-03", feature = "reqwest-010"))]
 pub type AsyncHttpClientError = Error<reqwest_0_10::Error>;
 
-#[cfg(feature = "reqwest-010")]
 mod blocking {
     use super::super::{HttpRequest, HttpResponse};
     use super::Error;
 
-    #[cfg(feature = "reqwest-010")]
     pub use reqwest_0_10 as reqwest;
-    #[cfg(feature = "reqwest-010")]
     use reqwest_0_10::blocking;
-    #[cfg(feature = "reqwest-010")]
     use reqwest_0_10::redirect::Policy as RedirectPolicy;
 
     use std::io::Read;
@@ -103,5 +94,57 @@ mod blocking {
     }
 }
 
-#[cfg(all(feature = "reqwest-010", feature = "futures-03"))]
-mod async_client;
+mod async_client {
+    use super::super::{HttpRequest, HttpResponse};
+    use super::Error;
+
+    pub use reqwest_0_10 as reqwest;
+    use reqwest_0_10::redirect::Policy as RediretPolicy;
+
+    use http::header::HeaderName;
+    use http::{HeaderMap, HeaderValue, StatusCode};
+
+    ///
+    /// Asynchronous HTTP client.
+    ///
+    pub async fn async_http_client(
+        request: HttpRequest,
+    ) -> Result<HttpResponse, Error<reqwest::Error>> {
+        let client = reqwest::Client::builder()
+            // Following redirects opens the client up to SSRF vulnerabilities.
+            .redirect(RediretPolicy::none())
+            .build()
+            .map_err(Error::Reqwest)?;
+
+        let mut request_builder = client
+            .request(request.method, request.url.as_str())
+            .body(request.body);
+        for (name, value) in &request.headers {
+            request_builder = request_builder.header(name.as_str(), value.as_bytes());
+        }
+        let request = request_builder.build().map_err(Error::Reqwest)?;
+
+        let response = client.execute(request).await.map_err(Error::Reqwest)?;
+
+        let status_code = response.status();
+        let headers = response
+            .headers()
+            .iter()
+            .map(|(name, value)| {
+                (
+                    HeaderName::from_bytes(name.as_str().as_ref())
+                        .expect("failed to convert HeaderName from http 0.2 to 0.1"),
+                    HeaderValue::from_bytes(value.as_bytes())
+                        .expect("failed to convert HeaderValue from http 0.2 to 0.1"),
+                )
+            })
+            .collect::<HeaderMap>();
+        let chunks = response.bytes().await.map_err(Error::Reqwest)?;
+        Ok(HttpResponse {
+            status_code: StatusCode::from_u16(status_code.as_u16())
+                .expect("failed to convert StatusCode from http 0.2 to 0.1"),
+            headers,
+            body: chunks.to_vec(),
+        })
+    }
+}
